hotelier-service изменение структуры, готов к проверке
