+++ README.md
@@ -3,98 +3,89 @@
 
 # Сруктура проекта
 ```bash
-final-project/
-├── api/                    # Папка для API контрактов
-│   ├── protobufs/          # Контракты API в формате Protocol Buffers (gRPC или REST)
-│   └── openapi/            # Контракты API в формате OpenAPI/Swagger
-│
-├── cmd/                    # Папка для запуска сервисов
-│   ├── booking-service/    # Основной сервис бронирования
-│   │   └── main.go         # Точка входа для сервиса бронирования
-│   ├── hotelier-service/   # Сервис для отельеров
-│   │   └── main.go         # Точка входа для сервиса отельеров
-│   └── client-service/     # Клиентский сервис
-│       └── main.go         # Точка входа для клиентского сервиса
-│
-├── configs/                # Конфигурации для всех сервисов
-│   ├── booking-config.yaml # Конфигурация сервиса бронирования
-│   ├── hotelier-config.yaml# Конфигурация сервиса отельеров
-│   └── client-config.yaml  # Конфигурация клиентского сервиса
-│
-├── deployments/            # Файлы для развертывания сервисов
-│   └── docker-compose.yml  # Docker Compose файл для всех сервисов и зависимостей
-│
-├── internal/               # Пакеты, которые не предназначены для использования вне проекта
-│   ├── booking/            # Пакеты сервиса бронирования
-│   ├── hotelier/           # Пакеты сервиса отельеров
-│   └── client/             # Пакеты клиентского сервиса
-│       └── repository/     # Логика работы с базой данных для клиентского сервиса
-│
-├── pkg/                    # Общие пакеты, которые можно переиспользовать между сервисами
-│   ├── logger/             # Пакет для структурированного логирования
-│   └── tracing/            # Пакет для трассировки
-│
-├── migrations/             # SQL-файлы для миграции базы данных
-│   ├── 001_create_tables.sql
-│   └── 002_add_indexes.sql
-│
-├── scripts/                # Скрипты для автоматизации работы с проектом
-│   ├── init_db.sh          # Скрипт инициализации базы данных
-│   └── run_tests.sh        # Скрипт для запуска всех тестов
-│
-├── test/                   # Папка для интеграционных тестов и моков
-│   ├── booking_tests/      # Интеграционные тесты сервиса бронирования
-│   └── mocks/              # Моки для тестирования
-│
-├── .env                    # Файл переменных окружения для настройки сервисов
-├── .gitignore              # Список файлов и папок, игнорируемых git
-└── README.md               # Основная документация по проекту
 
 
 
-
-
+hotel-service-go/
+├── booking-service/            # Сервис для управления бронированием, не реализова
+├── client-service/             # Сервис для управления данными клиентов, не реализован
+├── deployments/                # Конфигурации для деплоя
+│   └── docker-compose.yml      # Файл для настройки Docker Compose, для hotel-service
+├── hotel-service/              # Основной сервис управления отелями
+│   ├── api/                    # API спецификации
+│   │   ├── openapi/            # OpenAPI спецификации
+│   │   └── protobufs/          # Протобуф-контракты, нет
+│   ├── configs/                # Конфигурационные файлы сервиса
+│   │   └── hotelier-config.yaml # Основной конфигурационный файл
+│   ├── hotelier-service/       # Код основного сервиса
+│   │   ├── Dockerfile          # Dockerfile для сборки сервиса
+│   │   ├── main.go             # Точка входа сервиса
+│   │   └── service.go          # обработчик запросов
+│   ├── hotelier_tests/         # Тесты для сервиса
+│   │   └── service_test.go     # Тесты для модуля hotelier-service
+│   ├── internal/               # Внутренние модули и библиотеки
+│   │   └── hotelier/           # Реализация бизнес-логики сервиса
+│   │       └── service.go      # Бизнес-логика отеля
+│   ├── tracing/                # Настройки трассировки 
+│   └── metrics/                # Реализация метрик
+│       └── metrics.go          # потом сделаем
+├── migrations/                 # SQL-скрипты для управления базой данных
+│   ├── 001_create_tables.sql   # Скрипт для создания таблиц
+│   └── 002_add_indexes.sql     # Скрипт для добавления индексов не реализован
+├── test/                       # Тестирование
+│   ├── booking_tests/          # Тесты для сервиса бронирования не реализовано
+│   └── mocks/                  # Моки для тестов не реализовано
+├── .env                        # Переменные окружения
+├── go.mod                      # Модульный файл Go
+└── README.md                   # Описание проекта
 
 
 
 ```
 
 ### План реализации hotelier service
-#### 1. Контракты и API (api/openapi/hotelier.yaml):
+#### 1. Контракты и API:
 Контракт сервиса содержит эндпоинты для управления отелями и номерамb:
-{ 
-* POST /hotels: Создать отель
-* GET /hotels: Получить список отелей
-* POST /rooms: Создать номер
-* GET /rooms: Получить номера отеля
+{
+в hotelier.yaml подробнее
 }
 #### 2. Миграции (migrations/):
    SQL-файлы для создания таблиц Hotels и Rooms, а также их отношений.
-#### 3. Основная логика (internal/hotelier/):
+#### 3. Основная логика (hotel-service/internal/hotelier/):
    Основные функции, такие как добавление отеля и номера, получение доступных номеров.
-#### 4. Тесты (test/hotelier_tests/):
+#### 4. Тесты (hotel-service/test/hotelier_tests/):
    Тесты для проверки эндпоинтов и бизнес-логики.
 #### Конфигурация (configs/hotelier-config.yaml):
  Настройки подключения к базе данных, логирования и переменные окружения.
 
-### Подробнее о реализации:
-* API Контракты (api/openapi/hotelier.yaml). Создадим спецификацию API для отелей и номеров.
-* Миграции (migrations/001_create_tables.sql). SQL для создания таблиц Hotels и Rooms.
-* Основная логика (internal/hotelier/) Создадим основные функции сервиса для обработки отелей и номеров. internal/hotelier/service.go:
-* Тесты (test/hotelier_tests/service_test.go)
-* Docker и Конфигурация
-  1. docker-compose.yml Файл docker-compose.yml задает конфигурацию для базы данных PostgreSQL и сервиса hotelier.
-  2. Конфигурация Go сервиса (configs/hotelier-config.yaml) Конфигурационный файл для сервиса hotelier в формате YAML. В нем указываются параметры подключения к базе данных и порты.
-  3. Dockerfile для hotelier service Dockerfile для сборки и запуска сервиса hotelier.
-* Скрипты миграции:
-  Для автоматического применения миграций, добавить инициализацию базы данных в Dockerfile .
-```dockerfile
-COPY migrations /migrations
-RUN ./migrate -path /migrations -database "postgres://hotelier:hotelier_password@db:5432/hotelier_db?sslmode=disable" up
-```
+
 * Запуск
   Для запуска hotelier service и базы данных в Docker:
 ``` bash
 docker-compose up --build
 ```
-Этот набор конфигураций поднимет базу данных PostgreSQL, настроит соединение с hotelier service и откроет API на порту 8080.
+## Немного о запросах 
+
+Получение всех комнат:
+```bash
+curl -X GET http://localhost:8080/rooms
+```
+Получение комнат для конкретного отеля:
+```bash
+curl -X GET "http://localhost:8080/rooms?hotel_id=1"
+```
+
+Если вы пытаетесь создать комнату для отеля, который не существует:
+Запрос:
+```bash
+curl -X POST -H "Content-Type: application/json" \
+-d '{"hotel_id": 99, "room_number": "101", "price": 150.00}' \
+http://localhost:8080/rooms
+```
+Ответ (JSON):
+```json
+{
+"error": "hotel with the given ID does not exist"
+}
+```
+## подробнее в hotel-service/api/openapi
