--- cmd/booking-service/main.go
@@ -1,1 +1,0 @@
-package booking_service
