--- cmd/client-service/main.go
@@ -1,1 +1,0 @@
-package client_service
