+++ hotel-service/configs/hotelier-config.yaml
