+++ metrics/metrics.go
@@ -1,19 +1,20 @@
 package metrics
 
-import (
-	"github.com/prometheus/client_golang/prometheus"
-)
-
-var (
-	HttpRequestsTotal = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "http_requests_total",
-			Help: "Total number of HTTP requests",
-		},
-		[]string{"path", "method"},
-	)
-)
-
-func Init() {
-	prometheus.MustRegister(HttpRequestsTotal)
-}
+//
+//import (
+//	"github.com/prometheus/client_golang/prometheus"
+//)
+//
+//var (
+//	HttpRequestsTotal = prometheus.NewCounterVec(
+//		prometheus.CounterOpts{
+//			Name: "http_requests_total",
+//			Help: "Total number of HTTP requests",
+//		},
+//		[]string{"path", "method"},
+//	)
+//)
+//
+//func Init() {
+//	prometheus.MustRegister(HttpRequestsTotal)
+//}
